+++ shared/src/components/Button/Secondary/ButtonSecondarySmall.vue
@@ -14,6 +14,12 @@ export default {
   name: 'ButtonSecondarySmall',
   directives: {
     focus
+  },
+  props: {
+    disabled: {
+      type: Boolean,
+      default: false
+    }
   }
 }
 </script>