+++ shared/docs/v0.0.1/changelog.md
@@ -1,6 +1,8 @@
 # v0.0.1
 
 ## 🚀 Features:
+- Tag component created with tests and stories 
+- Sheet component created
 
 ## 🐛 Fixes: